move tc
